Cleanup caching - no model_name for caches - CleanThetaCache and CleanScoreCache apis

+++ src/artm/core/cache_manager.cc
@@ -33,37 +33,27 @@ void CacheManager::RequestMasterComponentInfo(MasterComponentInfo* master_info) 
       continue;
 
     MasterComponentInfo::CacheEntryInfo* info = master_info->add_cache_entry();
-    info->set_key(boost::lexical_cast<std::string>(key.first) + "|" + key.second);
+    info->set_key(boost::lexical_cast<std::string>(key));
     info->set_byte_size(entry->ByteSize());
   }
 }
 
-void CacheManager::DisposeModel(ModelName model_name) {
+void CacheManager::Clear() {
   auto keys = cache_.keys();
   for (auto &key : keys) {
     auto cache_entry = cache_.get(key);
-    if (cache_entry == nullptr) {
-      continue;
-    }
-
-    if (cache_entry->model_name() == model_name) {
-      if (cache_entry->has_filename()) {
-        try { fs::remove(fs::path(cache_entry->filename())); } catch(...) {}
-      }
-      cache_.erase(key);
+    if (cache_entry != nullptr && cache_entry->has_filename()) {
+      try { fs::remove(fs::path(cache_entry->filename())); } catch(...) {}
     }
   }
+
+  cache_.clear();
 }
 
 void CacheManager::RequestThetaMatrix(const GetThetaMatrixArgs& get_theta_args,
                                       ::artm::ThetaMatrix* theta_matrix) const {
-  std::string model_name = get_theta_args.model_name();
-  std::vector<CacheKey> keys = cache_.keys();
-
+  auto keys = cache_.keys();
   for (auto &key : keys) {
-    if (key.second != model_name)
-      continue;
-
     std::shared_ptr<DataLoaderCacheEntry> cache = cache_.get(key);
     if (cache == nullptr)
       continue;
@@ -75,17 +65,12 @@ void CacheManager::RequestThetaMatrix(const GetThetaMatrixArgs& get_theta_args,
     } else {
       BatchHelpers::PopulateThetaMatrixFromCacheEntry(*cache, get_theta_args, theta_matrix);
     }
-
-    if (get_theta_args.clean_cache()) {
-      cache_.erase(key);
-    }
   }
 }
 
 std::shared_ptr<DataLoaderCacheEntry> CacheManager::FindCacheEntry(
-    const boost::uuids::uuid& batch_uuid, const ModelName& model_name) const {
-  CacheKey key(batch_uuid, model_name);
-  std::shared_ptr<DataLoaderCacheEntry> retval = cache_.get(key);
+    const boost::uuids::uuid& batch_uuid) const {
+  std::shared_ptr<DataLoaderCacheEntry> retval = cache_.get(batch_uuid);
   if (retval == nullptr || !retval->has_filename())
     return retval;
 
@@ -103,10 +88,8 @@ std::shared_ptr<DataLoaderCacheEntry> CacheManager::FindCacheEntry(
 void CacheManager::UpdateCacheEntry(std::shared_ptr<DataLoaderCacheEntry> cache_entry) const {
   std::string uuid_str = cache_entry->batch_uuid();
   boost::uuids::uuid uuid(boost::uuids::string_generator()(uuid_str.c_str()));
-  ModelName model_name = cache_entry->model_name();
-  CacheKey cache_key(uuid, model_name);
-  std::shared_ptr<DataLoaderCacheEntry> old_entry = cache_.get(cache_key);
-  cache_.set(cache_key, cache_entry);
+  std::shared_ptr<DataLoaderCacheEntry> old_entry = cache_.get(uuid);
+  cache_.set(uuid, cache_entry);
   if (old_entry != nullptr && old_entry->has_filename()) {
     try { fs::remove(fs::path(old_entry->filename())); } catch(...) {}
   }
